--- src/components/common/Extensions/CalamityBanner/CalamityBanner.module.css.d.ts
@@ -1,12 +1,0 @@
-declare namespace CalamityBannerModuleCssNamespace {
-  export interface ICalamityBannerModuleCss {
-    calamityBannerInner: string;
-  }
-}
-
-declare const CalamityBannerModuleCssModule: CalamityBannerModuleCssNamespace.ICalamityBannerModuleCss & {
-  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
-  locals: CalamityBannerModuleCssNamespace.ICalamityBannerModuleCss;
-};
-
-export = CalamityBannerModuleCssModule;
