+++ src/components/v5/shared/CalamityBanner/hooks.ts
